fix(mail): refactor/reslove name collision in imports & rename submodules
