+++ src/__mail__/__init__.py
@@ -2,10 +2,10 @@
 Docstring for src.__mail__
 """
 
-from .welcome import welcome
-from .forgot_password import forgot_password
-from .email_verification import email_verification
-from .two_factor_authentication import two_factor_authentication
+from ._welcome import welcome
+from ._forgot_password import forgot_password
+from ._email_verification import email_verification
+from ._two_factor_authentication import two_factor_authentication
 
 __all__ = [
     "welcome",
